requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.3.0
tqdm>=4.62.0
tensorboard>=2.8.0
pyyaml>=5.4.0
h5py>=3.1.0
thop>=0.0.31
pillow>=8.3.0
optuna>=3.0.0